Update project strucutre
This commit contains a lot of changes so make sure everything is clear.

All exisitng sources and headers for the emulator project are moved to a separate directory. The main file emulator.c is still under src. The structure for the assembler project will be similar.

The Makefile should not require new modifications when more sources are added as it traverses the tree under each source folder.

Unit test support has been added (MinUnit). There is a snippet of the basic template for a unit test file. Running 'make test' compiles and runs all tests. Stick to the naming convention and the folder structure.