Implement new design for the arm11 memory
The memory is now an array of instructions. In this way, to write to or
read from an instruction we need only to deal with the new bin field in
an instruction, that represents the instruction in big endian. We also
changed all the functions to read from and write to memory, according to
this change. Started implementing tests.